Moved more stuff to render crate
