+++ deps/render/src/pipelines/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod background_pipeline;
+pub(crate) mod path_pipeline;
+pub(crate) mod polygon_pipeline;
+pub(crate) mod rect_pipeline;
