+++ deps/render/src/pipelines/rect_pipeline.rs
@@ -13,9 +13,13 @@ use wgpu::{
     Buffer, PipelineLayoutDescriptor, PrimitiveTopology, RenderPass, RenderPipeline, ShaderModuleDescriptor,
     ShaderSource, ShaderStages,
 };
-use window::{BufferUsages, DeviceHelper, PolygonMode, VecBuffer, VertexLayout, Window, image::Image};
+use window::{BufferUsages, DeviceHelper, PolygonMode, Window, image::Image};
 
-use crate::uniform::{UniformBind, make_uniform_layout};
+use crate::{
+    uniform::{UniformBind, make_uniform_layout},
+    vec_buffer::VecBuffer,
+    vertex_layout::VertexLayout,
+};
 
 const VERTICES: &[Point] = &[
     Point::new(-1.0, 1.0),
