+++ deps/render/src/pipelines/shaders/path.wgsl
