+++ deps/render/src/pipelines/shaders/polygon.wgsl
