+++ deps/render/src/pipelines/shaders/rect.wgsl
