+++ deps/render/src/pipelines/shaders/sprite_textured.wgsl
