+++ deps/render/src/pipelines/shaders/ui_image.wgsl
