+++ deps/window/src/lib.rs
@@ -1,11 +1,10 @@
 mod app;
-pub mod image;
 mod text;
 pub mod utils;
 mod window;
 
 mod frame_counter;
-mod render;
+pub mod image;
 mod screenshot;
 mod state;
 mod surface;
@@ -14,12 +13,6 @@ mod vertex_buffer;
 pub use app::*;
 pub use bytemuck::cast_slice;
 pub use image_proc::include_images;
-pub use render::{
-    path_data::PathData,
-    sprite_drawer::shader_data::{SpriteInstance, SpriteView},
-    vec_buffer::VecBuffer,
-    vertex_layout::VertexLayout,
-};
 pub use screenshot::*;
 pub use text::*;
 pub use utils::DeviceHelper;
