--- deps/window/src/render/mod.rs
@@ -1,4 +1,0 @@
-pub mod path_data;
-pub mod sprite_drawer;
-pub mod vec_buffer;
-pub mod vertex_layout;
