--- deps/window/src/render/sprite_drawer/mod.rs
@@ -1,1 +1,0 @@
-pub mod shader_data;
